Move remaining statistics/items/suggestions to their own modules for Holy Paladin parser
